# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Practical Test",
        "description": "For practice purposes",
        "version": "1"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tweets": {
            "get": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Index",
                "description": "List of tweets",
                "operationId": "indexTweet",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Store",
                "description": "Create new tweet",
                "operationId": "storeTweet",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "tweet",
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "tweet": {
                                        "type": "string",
                                        "example": "Sample Tweet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tweets/{id}": {
            "get": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Show",
                "description": "Show detail of tweet",
                "operationId": "showTweet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Delete",
                "description": "Tweets delete",
                "operationId": "deleteTweet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Update",
                "description": "Update tweet",
                "operationId": "updateTweet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tweet"
                                ],
                                "properties": {
                                    "tweet": {
                                        "type": "string",
                                        "example": "Sample Tweet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tweets/{id}/images": {
            "post": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Store Image",
                "description": "Create new tweet image",
                "operationId": "storeTweetImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tweets/{id}/images/{image}": {
            "delete": {
                "tags": [
                    "Tweets"
                ],
                "summary": "Tweets Delete Image",
                "description": "Delete new tweet image",
                "operationId": "deleteTweetImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}